[refactor] utils, types, constants 분리
- 가독성 개선을 위한 파일 분리
